+++ packages/@livestore/react/src/mod.ts
@@ -1,14 +1,12 @@
 export { StoreRegistry, storeOptions } from '@livestore/livestore'
 export { LiveList, type LiveListProps } from './experimental/components/LiveList.tsx'
 export * from './StoreRegistryContext.tsx'
-export {
-  type Dispatch,
-  type SetStateAction,
-  type SetStateActionPartial,
-  type StateSetters,
-  type UseClientDocumentResult,
-  useClientDocument,
+export type {
+  Dispatch,
+  SetStateAction,
+  SetStateActionPartial,
+  StateSetters,
+  UseClientDocumentResult,
 } from './useClientDocument.ts'
-export { useQuery, useQueryRef } from './useQuery.ts'
 export { type ReactApi, useStore, withReactApi } from './useStore.ts'
 export { useStackInfo } from './utils/stack-info.ts'